Moved LeftPaneNavigator to navigator package
