+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/common/di/WooPosActivityProvidesModule.kt
@@ -1,6 +1,6 @@
 package com.woocommerce.android.ui.woopos.common.di
 
-import com.woocommerce.android.ui.woopos.home.items.LeftPaneNavigator
+import com.woocommerce.android.ui.woopos.home.navigation.LeftPaneNavigator
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
