+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/home/navigation/WooPosLeftPaneScreensViewModel.kt
@@ -1,7 +1,6 @@
 package com.woocommerce.android.ui.woopos.home.navigation
 
 import androidx.lifecycle.ViewModel
-import com.woocommerce.android.ui.woopos.home.items.LeftPaneNavigator
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
